requirements.txt
pandas
numpy
streamlit
PyYAML
agraph-python